// src/core/actions/types-actions/profile-types.ts
export enum ProfileActionsType {
	START_LOAD_PROFILE = '[PROFILE] START_LOAD_PROFILE',
	SUCCESS_LOAD_PROFILE = '[PROFILE] SUCCESS_LOAD_PROFILE',
	ERROR_LOAD_PROFILE = '[PROFILE] ERROR_LOAD_PROFILE',

	START_UPLOAD_PHOTO_PROFILE = '[PROFILE] START_UPLOAD_PHOTO_PROFILE',
	SUCCESS_UPLOAD_PHOTO_PROFILE = '[PROFILE] SUCCESS_UPLOAD_PHOTO_PROFILE',
	ERROR_UPLOAD_PHOTO_PROFILE = 'ERROR_UPLOAD_PHOTO_PROFILE',

	START_DOWNLOAD_PHOTO_PROFILE = '[PROFILE] START_DOWNLOAD_PHOTO_PROFILE',
	SUCCESS_DOWNLOAD_PHOTO_PROFILE = '[PROFILE] SUCCESS_DOWNLOAD_PHOTO_PROFILE',
	ERROR_DOWNLOAD_PHOTO_PROFILE = '[PROFILE] ERROR_DOWNLOAD_PHOTO_PROFILE',

	START_UPDATE_NAME_PROFILE = '[PROFILE] START_UPDATE_NAME_PROFILE',
	SUCCESS_UPDATE_NAME_PROFILE = '[PROFILE] SUCCESS_UPDATE_NAME_PROFILE',
	ERROR_UPDATE_NAME_PROFILE = '[PROFILE] ERROR_UPDATE_NAME_PROFILE',

	START_UPDATE_SURNAME_PROFILE = '[PROFILE] START_UPDATE_SURNAME_PROFILE',
	SUCCESS_UPDATE_SURNAME_PROFILE = '[PROFILE] SUCCESS_UPDATE_SURNAME_PROFILE',
	ERROR_UPDATE_SURNAME_PROFILE = '[PROFILE] ERROR_UPDATE_SURNAME_PROFILE',
}